// src/api/types.gen.ts
export type JsonNode = {
    [key: string]: unknown;
};
// ...
export type Status = {
    /**
     * A mandatory connection status field.
     */
    connection?: 'CONNECTED' | 'DISCONNECTED' | 'STATELESS' | 'UNKNOWN' | 'ERROR';
    /**
     * The identifier of the object
     */
    id?: string;
    /**
     * The datetime of the last activity through this connection
     */
    lastActivity?: string;
    /**
     * A message associated with the state of a connection
     */
    message?: string;
    /**
     * A object status field.
     */
    runtime?: 'STARTED' | 'STOPPED';
    /**
     * The datetime the object was 'started' in the system.
     */
    startedAt?: string;
    /**
     * The type of the object
     */
    type?: string;
};
// ...
export type StatusList = {
    /**
     * List of result items that are returned by this endpoint
     */
    items: Array<Status>;
};
// ...
export type Adapter = {
    config?: JsonNode;
    /**
     * The adapter id, must be unique and only contain alpha numeric characters with spaces and hyphens.
     */
    id: string;
    status?: Status;
    /**
     * The adapter type associated with this instance
     */
    type?: string;
};
// ...
export type AdaptersList = {
    /**
     * List of result items that are returned by this endpoint
     */
    items: Array<Adapter>;
};
// ...
export type ProtocolAdapterCategory = {
    /**
     * The description associated with the category.
     */
    description?: string;
    /**
     * The display name of the category to be used in HCIs.
     */
    displayName: string;
    /**
     * The image associated with the category.
     */
    image?: string;
    /**
     * The unique name of the category to be used in API communication.
     */
    name: string;
};

/**
 * List of result items that are returned by this endpoint
 */
export type ProtocolAdapter = {
    /**
     * The author of the adapter
     */
    author?: string;
    /**
     * The capabilities of this adapter
     */
    capabilities?: Array<'READ' | 'DISCOVER' | 'WRITE' | 'COMBINE'>;
    category?: ProtocolAdapterCategory;
    configSchema?: JsonNode;
    /**
     * The description
     */
    description?: string;
    /**
     * The id assigned to the protocol adapter type
     */
    id?: string;
    /**
     * Is the adapter installed?
     */
    installed?: boolean;
    /**
     * The logo of the adapter
     */
    logoUrl?: string;
    /**
     * The name of the adapter
     */
    name?: string;
    /**
     * The supported protocol
     */
    protocol?: string;
    /**
     * The provisioning url of the adapter
     */
    provisioningUrl?: string;
    /**
     * The search tags associated with this adapter
     */
    tags?: Array<string>;
    uiSchema?: JsonNode;
    /**
     * The url of the adapter
     */
    url?: string;
    /**
     * The installed version of the adapter
     */
    version?: string;
};

export type ProtocolAdaptersList = {
    /**
     * List of result items that are returned by this endpoint
     */
    items: Array<ProtocolAdapter>;
};